requirements.txt
asgiref==3.5.2
Django==4.1
django-baton==2.4.1
django-dotenv==1.4.2
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
